Update configure's HTSlib detection and remove Samtools detection
Use Samtools's ax_with_htslib.m4 to get the same --with-htslib=DIR
configure option to build against either a source HTSlib directory
or an already-installed HTSlib. Produce an error for the previous
HTSLIB=DIR variable approach, though it would be possible to make
it equivalent instead if desired.

Remove the _USE_KNETFILE define, as it is ignored by HTSlib.