Fix vcf_var_has_info_key()'s key comparison
Previously any key would be returned as being matched by the empty
"key=value" at the start of ";HRUN=3" (which is invalid VCF).

This resulted in a crash when e.g. lofreq_filter.c looking for "SB"
expected a value but in this case the returned sb_char was NULL.
Crash fixed by no longer returning a hit for "SB" in this case.

(Also convert a couple of previously missed legacy samtools API usages.)